backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-firestore
google-cloud-storage
google-genai
python-multipart
google-auth
python-dotenv
google-auth-oauthlib